Simulate N-key autocommit with segmentation at a fixed code length.
git-svn-id: https://rimeime.googlecode.com/svn/trunk/ibus-rime@578 b90aed45-4297-9963-ff9d-98b604ac09a9

+++ Makefile
@@ -36,4 +36,4 @@ uninstall:
 	rm -R $(libexecdir)/ibus-rime
 
 clean:
-	rm -R build
+	if  [ -e build ]; then rm -R build; fi
